backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.1"
description = "Server for the Custom Agent Platform with Telegram integration"
authors = ["Cam Faith <c@encodex.dev>, Ewan May <e@encodex.dev>"]
[tool.poetry.dependencies]
python = "^3.9"
flask = "^3.0.3"
python-telegram-bot = "^21.6"
pymongo = "^4.10.1"
langchain = "^0.3.3"
langgraph = "^0.2.38"
python-dotenv = "^1.0.1"
flask-pymongo = "^2.3.0"
faker = "^30.6.0"
openai = "^1.51.2"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
pytest = "^8.3.3"
termcolor = "^2.5.0"
langchain-community = "^0.3.2"
langchain-core = "^0.3.10"
langchain-openai = "^0.2.2"
langchain-text-splitters = "^0.3.0"
langgraph-checkpoint = "^2.0.1"
langgraph-sdk = "^0.1.33"
gunicorn = "^21.2.0"
uvicorn = "^0.32.0"
fastapi = "^0.115.3"
langchain-anthropic = "^0.2.3"
langchain-google-vertexai = "^2.0.5"
langchain-ollama = "^0.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
black = "^24.10.0"
flake8 = "^7.1.1"
mongomock = "^4.2.0.post1"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
